+++ packages/react-native/android/src/main/kotlin/com/uulab/securekeypad/SecureKeypadView.kt
@@ -255,6 +255,7 @@ public open class SecureKeypadView @JvmOverloads constructor(
     override fun onAttachedToWindow() {
         super.onAttachedToWindow()
         requireSecureWindow()
+        requestSessionReconfigurationIfNeeded()
     }
 
     private fun requireSecureWindow() {